memmodel -> split
